Add backface culling

+++ src/boulder.h
@@ -2,6 +2,7 @@
 
 #include <SDL2/SDL.h>
 #include <SDL2_ttf/SDL_ttf.h>
+#include <cmath>
 #include <cstdlib>
 #include <sstream>
 
@@ -10,5 +11,5 @@ void Update(Uint32 deltaTime);
 void Render();
 void Shutdown();
 
-const int SCREEN_WIDTH = 1024;
-const int SCREEN_HEIGHT = 768;
+const float SCREEN_WIDTH = 1024.0;
+const float SCREEN_HEIGHT = 768.0;
